requirements.txt
google-cloud-pubsub==2.18.4
google-generativeai==0.3.1
python-dotenv==1.0.0
requests==2.31.0
